+++ CHANGELOG.md
@@ -5,13 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] -- 0000-00-00
+## [Unreleased]
 
 ### Added
 
 - Bugfix: Map text rendering not accounting for hi-res flag
 - Hook: TextRenderA
 - Hook: MapRenderB, MapRenderA
+- Core: Global State `race_state` `race_state_prev` `race_state_new`
+- Core: Ability to toggle Practice Mode OFF in race scene before countdown
+
+### Changed
+
+- Core: `PLUGIN_VERSION` - `17`→`18`
+
+### Removed
+
+- Core: Global State `player.in_race_count` `player.in_race_results` `player.in_race_racing`
 
 ### Fixed